+++ src/content/init-discovery.js
@@ -75,12 +75,70 @@ export function initDiscovery(options, data) {
 
     settingsPage(discovery);
 
-    discovery.page.define('default', [
-        {
-            view: 'struct',
-            expanded: '=+#.settings.expandLevel'
+    const btnCopyToClipboard = {
+        view: 'button',
+        content: 'text:"Copy to clipboard"',
+        onClick(_, { json }) {
+            copyToClipboard(json);
         }
-    ]);
+    };
+    const btnDownload = {
+        view: 'button',
+        content: 'text:"Download as file"',
+        onClick(_, { json }) {
+            downloadAsFile(json);
+        }
+    };
+
+    discovery.page.define('default', {
+        view: 'context',
+        modifiers: {
+            view: 'page-header',
+            content: [
+                btnCopyToClipboard,
+                btnDownload,
+                {
+                    view: 'block',
+                    content: [
+                        {
+                            view: 'button',
+                            className: 'collapse-all',
+                            content: 'text:"-"',
+                            onClick(el, data, { onChange }) {
+                                onChange(1, 'expandLevel');
+                            },
+                            postRender(el, config, data, context) {
+                                context.onChange = config.onChange;
+                                el.title = 'Collapse all';
+                                el.innerHTML = '<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" viewBox="0 0 16 16">' +
+                                    '<path fill-rule="evenodd" d="M1 8a.5.5 0 0 1 .5-.5h13a.5.5 0 0 1 0 1h-13A.5.5 0 0 1 1 8zm7-8a.5.5 0 0 1 .5.5v3.793l1.146-1.147a.5.5 0 0 1 .708.708l-2 2a.5.5 0 0 1-.708 0l-2-2a.5.5 0 1 1 .708-.708L7.5 4.293V.5A.5.5 0 0 1 8 0zm-.5 11.707-1.146 1.147a.5.5 0 0 1-.708-.708l2-2a.5.5 0 0 1 .708 0l2 2a.5.5 0 0 1-.708.708L8.5 11.707V15.5a.5.5 0 0 1-1 0v-3.793z"/>' +
+                                    '</svg>';
+                            }
+                        },
+                        {
+                            view: 'button',
+                            className: 'expand-all',
+                            content: 'text:"+"',
+                            onClick(el, data, { onChange }) {
+                                onChange(100, 'expandLevel');
+                            },
+                            postRender(el, config, data, context) {
+                                context.onChange = config.onChange;
+                                el.title = 'Expand all';
+                                el.innerHTML = '<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" viewBox="0 0 16 16">' +
+                                    '<path fill-rule="evenodd" d="M1 8a.5.5 0 0 1 .5-.5h13a.5.5 0 0 1 0 1h-13A.5.5 0 0 1 1 8zM7.646.146a.5.5 0 0 1 .708 0l2 2a.5.5 0 0 1-.708.708L8.5 1.707V5.5a.5.5 0 0 1-1 0V1.707L6.354 2.854a.5.5 0 1 1-.708-.708l2-2zM8 10a.5.5 0 0 1 .5.5v3.793l1.146-1.147a.5.5 0 0 1 .708.708l-2 2a.5.5 0 0 1-.708 0l-2-2a.5.5 0 0 1 .708-.708L7.5 14.293V10.5A.5.5 0 0 1 8 10z"/>' +
+                                    '</svg>';
+                            }
+                        }
+                    ]
+                }
+            ]
+        },
+        content: {
+            view: 'struct',
+            expanded: '=+(#.expandLevel or #.settings.expandLevel)'
+        }
+    });
 
     discovery.view.define('raw', (el, _, raw) => {
         const contentEl = el.appendChild(document.createElement('pre'));
@@ -93,7 +151,7 @@ export function initDiscovery(options, data) {
                 view: 'alert-warning',
                 className: 'too-big-json',
                 content: [
-                    'text:`JSON is too big (${size.weight()} bytes), only first ${firstSlice.size().weight()} is shown. Output the entire JSON may cause to browser freezing for a while. `',
+                    'text:`JSON is too big (${size.weight()} bytes), only first ${firstSlice.size().weight()} is shown. Output the entire JSON may cause to browser\'s tab freezing for a while. `',
                     {
                         view: 'button',
                         content: 'text:"Show all"',
@@ -120,20 +178,8 @@ export function initDiscovery(options, data) {
             {
                 view: 'page-header',
                 content: [
-                    {
-                        view: 'button',
-                        content: 'text:"Copy to clipboard"',
-                        onClick(_, { json }) {
-                            copyToClipboard(json);
-                        }
-                    },
-                    {
-                        view: 'button',
-                        content: 'text:"Download as file"',
-                        onClick(_, { json }) {
-                            downloadAsFile(json);
-                        }
-                    }
+                    btnCopyToClipboard,
+                    btnDownload
                 ]
             },
             'raw'