+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Improved data loading and page chages rollback when data is not a JSON or broken JSON
 - Added progressbar on data loading
 - Added inspect button to navbar
+- Rearranged navbar to be more compact
 
 ## 1.10.3 (01-04-2021)